More efficient window implementation (#217)
* Add more optimization rules, for example one that splits OVER from other projects

* Implemented a more optimized window handler plugin

* Huge refactoring to integrate multiple window aggregations into one pass - needs refinement still

* Make sure to keep output names even after optimization

* Refine the new implemenation with docu and type annotations

* Make sure the mapped function is pickleable without dask
