+++ dask_sql/physical/rex/core/__init__.py
@@ -1,6 +1,5 @@
 from .call import RexCallPlugin
 from .input_ref import RexInputRefPlugin
 from .literal import RexLiteralPlugin
-from .over import RexOverPlugin
 
-__all__ = [RexCallPlugin, RexInputRefPlugin, RexLiteralPlugin, RexOverPlugin]
+__all__ = [RexCallPlugin, RexInputRefPlugin, RexLiteralPlugin]
